Refactor some code, add flake ignore to init
