+++ DashAI/back/tasks/__init__.py
@@ -1,3 +1,4 @@
+# flake8: noqa
 from DashAI.back.tasks.base_task import BaseTask
 from DashAI.back.tasks.tabular_classification_task import TabularClassificationTask
 from DashAI.back.tasks.text_classification_task import TextClassificationTask
