+++ tests/back/registries/test_mapping_mixin.py
@@ -1,7 +1,7 @@
 import pytest
 
 from DashAI.back.registries.base_registry import BaseRegistry
-from DashAI.back.registries.mixins import TaskComponentMappingMixin
+from DashAI.back.registries.register_mixins import TaskComponentMappingMixin
 from DashAI.back.tasks import BaseTask
